Fixed Bug where MvxCommand Parameter could not be parsed / casted, when the CommandParameter was defined in XAML on WP8 and WinRT.
