+++ CrossCore/Cirrious.CrossCore/Cirrious.CrossCore.csproj
@@ -47,6 +47,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="BindingFlags.cs" />
+    <Compile Include="ExtensionMethods\MvxCrossCoreExtensions.cs" />
     <Compile Include="Converters\MvxBindingConstant.cs" />
     <Compile Include="Core\MvxDelegateExtensionMethods.cs" />
     <Compile Include="Core\MvxObjectExtensions.cs" />
