+++ CrossCore/Cirrious.CrossCore/ExtensionMethods/MvxCrossCoreExtensions.cs
@@ -1,0 +1,91 @@
+﻿// MvxBindingExtensions.cs
+// (c) Copyright Cirrious Ltd. http://www.cirrious.com
+// MvvmCross is licensed using Microsoft Public License (Ms-PL)
+// Contributions and inspirations noted in readme.md and license.txt
+// 
+// Project Lead - Stuart Lodge, @slodge, me@slodge.com
+
+using System;
+using System.Globalization;
+using System.Reflection;
+using Cirrious.CrossCore;
+using Cirrious.CrossCore.IoC;
+
+namespace Cirrious.CrossCore.ExtensionMethods
+{
+    public static class MvxCrossCoreExtensions
+    {
+        // core implementation of ConvertToBoolean
+        public static bool ConvertToBooleanCore(this object result)
+        {
+            if (result == null)
+                return false;
+
+            if (result is string)
+                return !string.IsNullOrEmpty((string)result);
+
+            if (result is bool)
+                return (bool)result;
+
+            var resultType = result.GetType();
+            if (resultType.GetTypeInfo().IsValueType)
+            {
+                var underlyingType = Nullable.GetUnderlyingType(resultType) ?? resultType;
+                return !result.Equals(underlyingType.CreateDefault());
+            }
+
+            return true;
+        }
+
+        // core implementation of MakeSafeValue
+        public static object MakeSafeValueCore(this Type propertyType, object value)
+        {
+            if (value == null)
+            {
+                return propertyType.CreateDefault();
+            }
+
+            var safeValue = value;
+            if (!propertyType.IsInstanceOfType(value))
+            {
+                if (propertyType == typeof(string))
+                {
+                    safeValue = value.ToString();
+                }
+                else if (propertyType.GetTypeInfo().IsEnum)
+                {
+                    if (value is string)
+                        safeValue = Enum.Parse(propertyType, (string)value, true);
+                    else
+                        safeValue = Enum.ToObject(propertyType, value);
+                }
+                else if (propertyType.GetTypeInfo().IsValueType)
+                {
+                    var underlyingType = Nullable.GetUnderlyingType(propertyType) ?? propertyType;
+                    if (underlyingType == typeof(bool))
+                        safeValue = value.ConvertToBooleanCore();
+                    else
+                        safeValue = ErrorMaskedConvert(value, underlyingType, CultureInfo.CurrentUICulture);
+                }
+                else
+                {
+                    safeValue = ErrorMaskedConvert(value, propertyType, CultureInfo.CurrentUICulture);
+                }
+            }
+            return safeValue;
+        }
+
+        private static object ErrorMaskedConvert(object value, Type type, CultureInfo cultureInfo)
+        {
+            try
+            {
+                return Convert.ChangeType(value, type, cultureInfo);
+            }
+            catch (Exception)
+            {
+                // pokemon - mask the error
+                return value;
+            }
+        }
+    }
+}
